Address review: share the drag-surface props between the home cards
gang-card.tsx and campaign-card.tsx each carried a verbatim copy of the
drag-surface style object and the contextmenu gate. Move both into
dragSurfaceProps() in use-dnd-sensors.ts, next to useSuppressClickAfterDrag,
so the workaround has one home and one explanatory comment.

No new file, and no behaviour change: the same style and handler are applied
to the same element, still only when the card is draggable. Re-ran the
Chromium checks — long-press context menu cancelled on favourite cards but not
plain ones, desktop right-click still allowed, same-position drop and
drag-and-return still don't navigate, touch and mouse reorder, plain taps,
favourite star and keyboard reorder all unchanged.

Named as a plain function rather than a hook since it calls none.
